requirements.txt
elasticsearch
python-dotenv
py7zr
minio
